Implemented changes requested by mde168
Streamlined code logic in the event of a draw
New Pseudocode to represent the changes made to code
New Flowchart to represent the changes made to code
Additional commentary added to Game Stage (now Game Stage (Round Manager)) to explain purpose of the new "early" termination check.
Re-organized the input stage section to discuss card descriptors before showing a full input as requested
Added a small discussion on the usage of cin.ignore() and its significance in the input code.
A change was not made to the section breakdown, I believe that the current section breakdown better encompasses the implementation produced in code over the separated stages discussed in the "real-life" game description.
cout changed to cerr for the unknown output state as suggested (oversight in implementation)