Implemented changes requested by mde168
Streamlined code logic in the event of a draw (as suggested)

Renamed input variables to be more descriptive of what they are (they were previously using the default variable names provided by CodinGame, they are now named as suggested)

Changed the default output behavior when gameWinner is not a defined value to use cerr instead of cout. 

New Pseudocode to represent the changes made to code

New Flowchart to represent the changes made to code

Additional commentary added to Game Stage (now Game Stage (Round Manager)) to explain purpose of the new "early" termination check.

Re-organized the input stage section to discuss card descriptors before showing a full input as requested
Added a small discussion on the usage of cin.ignore() and its significance in the input code.

A change was not made to the section breakdown, I believe that the current section breakdown better encompasses the implementation produced in code over the separated stages discussed in the "real-life" game description. This was discussed with Kenzie and we came to an agreeance on this factor.

Additional commentary added to explain what the "game-winner" is in the Simulating the Game section.

Additional commentary added to better explain the drawing mechanics of both the Fight/Battle stage and War stage in the War section (provided an example).